+++ Woof/Woof/Model/Walking/Rating.swift
@@ -1,13 +1,10 @@
-import Foundation
-
 /**
  The ratings given by both the pet owner and the pet sitter
- Conforms to the `Codable` protocols.
  */
-struct Ratings: Codable {
+struct Rating: Codable {
     /// The rating is given to walking by the pet owner.
-    let ownerRating: Int
+    let ownerRating: Double
 
     /// The rating is given to walking by the pet sitter.
-    let sitterRating: Int
+    let sitterRating: Double
 }