+++ Woof/Woof/Model/Walking/WalkingStatus.swift
@@ -1,43 +1,22 @@
-import Foundation
-
 /**
  The possible walking status values for a `Walking`.
- Conforms to the `Codable` protocol.
  */
-enum WalkingStatus: Codable {
+enum WalkingStatus: String, Codable {
     /// The walking is available for booking.
-    case available
+    case available = "Available"
 
     /// The walking is pending approval from the pet sitter.
-    case pending
+    case pending = "Pending"
 
-    /// The walking is booked by owner.
-    case booked
+    /// The walking has been booked by owner.
+    case booked = "Booked"
 
     /// The walking has been completed.
-    case finished
+    case finished = "Done"
 
     /// The walking has been canceled by the sitter.
-    case canceledBySitter
+    case canceledBySitter = "Cancelled"
 
     /// The walking has expired due to the elapsed duration since it was initially declared.
-    case expired
-
-    /// Text description of possible status.
-    var description: String {
-        switch self {
-        case .available:
-            return "Available"
-        case .pending:
-            return "Pending"
-        case .booked:
-            return "Booked"
-        case .finished:
-            return "Done"
-        case .canceledBySitter:
-            return "Cancelled"
-        case .expired:
-            return "Expired"
-        }
-    }
+    case expired = "Expired"
 }