+++ Woof/WoofTests/Model/Walking/WalkingStatusTests.swift
@@ -3,11 +3,11 @@ import XCTest
 final class WalkingStatusTests: XCTestCase {
     func testDescriptionReturnsExpectedStringAndExist() {
         // Given // When // Then
-        XCTAssertEqual(WalkingStatus.available.description, "Available")
-        XCTAssertEqual(WalkingStatus.pending.description, "Pending")
-        XCTAssertEqual(WalkingStatus.booked.description, "Booked")
-        XCTAssertEqual(WalkingStatus.finished.description, "Done")
-        XCTAssertEqual(WalkingStatus.canceledBySitter.description, "Cancelled")
-        XCTAssertEqual(WalkingStatus.expired.description, "Expired")
+        XCTAssertEqual(WalkingStatus.available.rawValue, "Available")
+        XCTAssertEqual(WalkingStatus.pending.rawValue, "Pending")
+        XCTAssertEqual(WalkingStatus.booked.rawValue, "Booked")
+        XCTAssertEqual(WalkingStatus.finished.rawValue, "Done")
+        XCTAssertEqual(WalkingStatus.canceledBySitter.rawValue, "Cancelled")
+        XCTAssertEqual(WalkingStatus.expired.rawValue, "Expired")
     }
 }