+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added 
 
-- Added `Walking` enity - [#88](https://github.com/IronFoundation/iOS-Woof/pull/88)
+- Add `Walking` entity - [#88](https://github.com/IronFoundation/iOS-Woof/pull/88)
 
 ## [0.5.0] - 27.12.2023